IMPROVE: make all user routes be in user router
