+++ src/server/user.ts
@@ -1,11 +1,20 @@
 import { XUtils } from "@vex-chat/crypto";
+import { XTypes } from "@vex-chat/types";
 import express from "express";
 import winston from "winston";
+import nacl from "tweetnacl";
+import { stringify } from "uuid";
 
 import { Database, hashPassword } from "../Database";
 import { censorUser } from "./utils";
 
-export const getUserRouter = (db: Database, log: winston.Logger) => {
+const TokenScopes = XTypes.HTTP.TokenScopes;
+
+export const getUserRouter = (
+    db: Database,
+    log: winston.Logger,
+    tokenValidator: (key: string, scope: XTypes.HTTP.TokenScopes) => boolean
+) => {
     const router = express.Router();
 
     router.get("/:id", async (req, res) => {
@@ -78,6 +87,46 @@ export const getUserRouter = (db: Database, log: winston.Logger) => {
             res.sendStatus(200);
         } catch (err) {
             res.sendStatus(500);
+        }
+    });
+
+    router.post("/:id/devices", async (req, res) => {
+        const devicePayload: XTypes.HTTP.IDevicePayload = req.body;
+
+        const userEntry = await db.retrieveUser(req.params.id);
+
+        if (!userEntry) {
+            res.sendStatus(404);
+            log.warn("User does not exist.");
+            return;
+        }
+
+        const salt = XUtils.decodeHex(userEntry.passwordSalt);
+        const payloadHash = XUtils.encodeHex(
+            hashPassword(devicePayload.password, salt)
+        );
+
+        if (payloadHash !== userEntry.passwordHash) {
+            res.sendStatus(401);
+            return;
+        }
+
+        const token = nacl.sign.open(
+            XUtils.decodeHex(devicePayload.signed),
+            XUtils.decodeHex(devicePayload.signKey)
+        );
+
+        if (!token) {
+            log.warn("Invalid signature on token.");
+            res.sendStatus(400);
+            return;
+        }
+
+        if (tokenValidator(stringify(token), TokenScopes.Device)) {
+            await db.createDevice(userEntry.userID, devicePayload);
+            res.sendStatus(200);
+        } else {
+            res.sendStatus(401);
         }
     });
